--- a/Cargo.toml
+++ b/Cargo.toml
@@ -86,3 +86,5 @@ codegen-units = 1
 # TODO: Remove once chacha20poly1305 0.10 appears in libp2p's dependencies
 chacha20poly1305 = { git = "https://github.com/RustCrypto/AEADs", rev = "06dbfb5571687fd1bbe9d3c9b2193a1ba17f8e99" }
 libp2p = { git = "https://github.com/subspace/rust-libp2p", branch = "subspace-v3" }
+# TODO: Remove once new version of parity-db will be released
+parity-db = { git = "https://github.com/paritytech/parity-db/", rev = "5c6abf17e0027935de9c78959c73b8c35835cc10" }