--- a/crates/subspace-farmer/src/commitments.rs
+++ b/crates/subspace-farmer/src/commitments.rs
@@ -7,9 +7,9 @@ use crate::plot::{PieceOffset, Plot};
 use arc_swap::ArcSwapOption;
 use databases::{CommitmentDatabases, CreateDbEntryResult, DbEntry};
 use event_listener_primitives::{Bag, HandlerId};
+use parity_db::Db;
 use parking_lot::Mutex;
 use rayon::prelude::*;
-use rocksdb::{WriteBatch, DB};
 use std::path::PathBuf;
 use std::sync::atomic::{AtomicBool, Ordering};
 use std::sync::Arc;
@@ -28,11 +28,13 @@ const TAGS_WRITE_BATCH_SIZE: usize = 16 * 1024 * 1024 / (TAG_SIZE + PIECE_OFFSET
 #[derive(Debug, Error)]
 pub enum CommitmentError {
     #[error("Metadata DB error: {0}")]
-    MetadataDb(rocksdb::Error),
+    MetadataDb(parity_db::Error),
     #[error("Commitment DB error: {0}")]
-    CommitmentDb(rocksdb::Error),
+    CommitmentDb(parity_db::Error),
     #[error("Plot error: {0}")]
     Plot(io::Error),
+    #[error("Migration error: {0}")]
+    Migrate(io::Error),
 }
 
 #[derive(Debug, Copy, Clone)]
@@ -132,7 +134,8 @@ impl Commitments {
 
             let db_path = self.inner.base_directory.join(hex::encode(salt));
             db_entry.lock().replace(Arc::new(
-                DB::open_default(db_path).map_err(CommitmentError::CommitmentDb)?,
+                Db::open_or_create(&CommitmentDatabases::options(db_path))
+                    .map_err(CommitmentError::CommitmentDb)?,
             ));
         }
 
@@ -173,11 +176,12 @@ impl Commitments {
                 if tags_with_offset.len() == tags_with_offset.capacity() {
                     tags_with_offset.sort_by(|(tag_a, _), (tag_b, _)| tag_a.cmp(tag_b));
 
-                    let mut batch = WriteBatch::default();
-                    for (tag, offset) in &tags_with_offset {
-                        batch.put(tag, offset);
-                    }
-                    db.write(batch).map_err(CommitmentError::CommitmentDb)?;
+                    db.commit(
+                        tags_with_offset
+                            .iter()
+                            .map(|(tag, offset)| (0, tag, Some(offset.to_vec()))),
+                    )
+                    .map_err(CommitmentError::CommitmentDb)?;
 
                     tags_with_offset.clear();
                 }
@@ -195,11 +199,12 @@ impl Commitments {
                 if let Some(db) = db_guard.as_ref() {
                     tags_with_offset.sort_by(|(tag_a, _), (tag_b, _)| tag_a.cmp(tag_b));
 
-                    let mut batch = WriteBatch::default();
-                    for (tag, offset) in &tags_with_offset {
-                        batch.put(tag, offset);
-                    }
-                    db.write(batch).map_err(CommitmentError::CommitmentDb)?;
+                    db.commit(
+                        tags_with_offset
+                            .iter()
+                            .map(|(tag, offset)| (0, tag, Some(offset.to_vec()))),
+                    )
+                    .map_err(CommitmentError::CommitmentDb)?;
                 }
             }
 
@@ -240,12 +245,12 @@ impl Commitments {
             let db_guard = db_entry.lock();
 
             if let Some(db) = db_guard.as_ref() {
-                let mut batch = WriteBatch::default();
-                for piece in pieces {
-                    let tag = create_tag(piece, salt);
-                    batch.delete(tag);
-                }
-                db.write(batch).map_err(CommitmentError::CommitmentDb)?;
+                db.commit(
+                    pieces
+                        .iter()
+                        .map(|piece| (0, create_tag(piece, salt), None)),
+                )
+                .map_err(CommitmentError::CommitmentDb)?;
             }
         }
 
@@ -276,11 +281,12 @@ impl Commitments {
 
                 tags_with_offset.sort_by(|(tag_a, _), (tag_b, _)| tag_a.cmp(tag_b));
 
-                let mut batch = WriteBatch::default();
-                for (tag, piece_offset) in tags_with_offset {
-                    batch.put(tag, piece_offset.to_le_bytes());
-                }
-                db.write(batch).map_err(CommitmentError::CommitmentDb)?;
+                db.commit(
+                    tags_with_offset
+                        .into_iter()
+                        .map(|(tag, offset)| (0, tag, Some(offset.to_le_bytes().to_vec()))),
+                )
+                .map_err(CommitmentError::CommitmentDb)?;
             };
         }
 
@@ -315,12 +321,17 @@ impl Commitments {
                 return Vec::new();
             }
         };
-        let iter = db.raw_iterator();
+        let iter = match db.iter(0) {
+            Ok(iter) => iter,
+            Err(_) => {
+                return Vec::new();
+            }
+        };
 
         // Take the best out of 10 solutions
         let mut solutions = SolutionIterator::new(iter, target, range)
-            .take(limit)
-            .collect::<Vec<_>>();
+            .map(|iter| iter.take(limit).collect::<Vec<_>>())
+            .unwrap_or_default();
         let target = u64::from_be_bytes(target);
         solutions.sort_by_key(|(tag, _)| {
             let tag = u64::from_be_bytes(*tag);
@@ -374,7 +385,7 @@ enum SolutionIteratorState {
 }
 
 pub(crate) struct SolutionIterator<'a> {
-    iter: rocksdb::DBRawIterator<'a>,
+    iter: parity_db::BTreeIterator<'a>,
     state: SolutionIteratorState,
     /// Lower bound of solution range
     lower: u64,
@@ -383,7 +394,11 @@ pub(crate) struct SolutionIterator<'a> {
 }
 
 impl<'a> SolutionIterator<'a> {
-    pub fn new(mut iter: rocksdb::DBRawIterator<'a>, target: Tag, range: u64) -> Self {
+    pub fn new(
+        mut iter: parity_db::BTreeIterator<'a>,
+        target: Tag,
+        range: u64,
+    ) -> parity_db::Result<Self> {
         let (lower, is_lower_overflowed) = u64::from_be_bytes(target).overflowing_sub(range / 2);
         let (upper, is_upper_overflowed) = u64::from_be_bytes(target).overflowing_add(range / 2);
 
@@ -394,29 +409,27 @@ impl<'a> SolutionIterator<'a> {
         );
 
         let state = if is_lower_overflowed || is_upper_overflowed {
-            iter.seek_to_first();
+            iter.seek_to_first()?;
             SolutionIteratorState::OverflowStart
         } else {
-            iter.seek(lower.to_be_bytes());
+            iter.seek(&lower.to_be_bytes())?;
             SolutionIteratorState::NoOverflow
         };
-        Self {
+        Ok(Self {
             iter,
             state,
             lower,
             upper,
-        }
+        })
     }
 
     fn next_entry(&mut self) -> Option<(Tag, PieceOffset)> {
-        self.iter
-            .key()
-            .map(|tag| tag.try_into().unwrap())
-            .map(|tag| {
-                let offset = u64::from_le_bytes(self.iter.value().unwrap().try_into().unwrap());
-                self.iter.next();
-                (tag, offset)
-            })
+        self.iter.next().ok().flatten().map(|(tag, offset)| {
+            (
+                tag.try_into().unwrap(),
+                u64::from_le_bytes(offset.try_into().unwrap()),
+            )
+        })
     }
 }
 
@@ -433,7 +446,7 @@ impl<'a> Iterator for SolutionIterator<'a> {
                 .filter(|(tag, _)| u64::from_be_bytes(*tag) <= self.upper)
                 .or_else(|| {
                     self.state = SolutionIteratorState::OverflowEnd;
-                    self.iter.seek(self.lower.to_be_bytes());
+                    self.iter.seek(&self.lower.to_be_bytes()).ok()?;
                     self.next()
                 }),
             SolutionIteratorState::OverflowEnd => self.next_entry(),