--- a/crates/subspace-farmer/src/commitments/metadata.rs
+++ b/crates/subspace-farmer/src/commitments/metadata.rs
@@ -1,11 +1,9 @@
 use crate::commitments::CommitmentError;
-use rocksdb::DB;
+use parity_db::Db;
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 use std::path::PathBuf;
 use subspace_core_primitives::Salt;
-
-const COMMITMENTS_KEY: &[u8] = b"commitments";
 
 #[derive(Debug, Copy, Clone, Eq, PartialEq, Serialize, Deserialize)]
 pub(super) enum CommitmentStatus {
@@ -15,15 +13,21 @@ pub(super) enum CommitmentStatus {
     Created,
 }
 
-#[derive(Debug)]
 pub(super) struct CommitmentMetadata {
-    db: DB,
+    db: Db,
+}
+
+impl std::fmt::Debug for CommitmentMetadata {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.debug_struct("CommitmentMetadata").finish_non_exhaustive()
+    }
 }
 
 impl CommitmentMetadata {
-    pub(super) fn new(path: PathBuf) -> Result<Self, CommitmentError> {
-        let db = DB::open_default(path).map_err(CommitmentError::MetadataDb)?;
+    const COMMITMENTS_KEY: &[u8] = b"commitments";
 
+    pub(super) fn new(path: PathBuf) -> Result<Self, CommitmentError> {
+        let db = Db::with_columns(&path, 1).map_err(CommitmentError::MetadataDb)?;
         Ok(Self { db })
     }
 
@@ -33,7 +37,7 @@ impl CommitmentMetadata {
     {
         let mut metadata = self
             .db
-            .get(COMMITMENTS_KEY)
+            .get(0, Self::COMMITMENTS_KEY)
             .map_err(CommitmentError::MetadataDb)?
             .map(|bytes| {
                 serde_json::from_slice::<HashMap<String, CommitmentStatus>>(&bytes)
@@ -52,7 +56,11 @@ impl CommitmentMetadata {
             .collect();
 
         self.db
-            .put(COMMITMENTS_KEY, &serde_json::to_vec(&metadata).unwrap())
+            .commit(std::iter::once((
+                0,
+                Self::COMMITMENTS_KEY,
+                Some(serde_json::to_vec(&metadata).unwrap()),
+            )))
             .map_err(CommitmentError::MetadataDb)
     }
 }