Fixes #567 

This pr migrates to parity db from rocksdb. I decided to migrate only piece index hash db (copying data over to parity db) and just remove all commitments (as we can regenerate them, and it shouldn't take much time).

### Code contributor checklist:
* [x] I have read, understood and followed [contributing guide](../CONTRIBUTING.md)