@i1i1 2022-08-27
Here is the [issue](https://github.com/paritytech/parity-db/issues/89) related to the release of `parity-db`