@i1i1 2022-08-29
Benchmarked changes, and it seems that performance has decreased:

Before:
```sh
$ cargo run --release -p subspace-farmer -- --base-path ... bench --plot-size 3G --write-pieces-size 6G
Finished benchmarking.

3.13G allocated for farming
2.79G actual space pledged (which is 89.25%)
344.44M of overhead (which is 10.75%)
1m 26s plotting time
34.69M/s average plotting throughput
Recommitment took 1.239412406s
```

After:
```sh
$ cargo run --release -p subspace-farmer -- --base-path ... bench --plot-size 3G --write-pieces-size 6G
Finished benchmarking.

3.20G allocated for farming
2.79G actual space pledged (which is 87.31%)
415.85M of overhead (which is 12.69%)
2m 24s plotting time
20.72M/s average plotting throughput
Recommitment took 1.071120161s