@nazar-pc 2022-08-30
I'd wait for upstream PR to be accepted though or at least address their concerns as we can't afford have stack overflows.